+++ tgui/packages/tgui/interfaces/LootPanel/IconDisplay.tsx
@@ -18,14 +18,14 @@ export function IconDisplay(props: Props) {
     size: { height, width },
   } = props;
 
-  const fallback = <Icon name="spinner" size={2.2} spin color="gray" />;
+  const fallback = <Icon name="spinner" size={1.5} spin color="gray" />;
 
   if (!icon) {
     return fallback;
   }
 
   if (icon === 'n/a') {
-    return <Icon name="dumpster-fire" size={2} color="gray" />;
+    return <Icon name="dumpster-fire" size={1.5} color="gray" />;
   }
 
   if (icon_state) {
