+++ tgui/packages/tgui/interfaces/LootPanel/index.tsx
@@ -1,4 +1,5 @@
 import { useState } from 'react';
+import { useMemo } from 'react';
 import { Button, Input, Section, Stack } from 'tgui-core/components';
 import { isEscape } from 'tgui-core/keys';
 import { clamp } from 'tgui-core/math';
@@ -19,12 +20,40 @@ export function LootPanel(props) {
   const { act, data } = useBackend<Data>();
   const { contents = [], searching } = data;
 
-  const [grouping, setGrouping] = useState(false);
+  // limitations: items with different stack counts, charges etc.
+  const contentsByPath = useMemo(() => {
+    const acc: Record<string, SearchItem[]> = {};
+
+    for (let i = 0; i < contents.length; i++) {
+      const item = contents[i];
+      if (item.path) {
+        if (!acc[item.path]) {
+          acc[item.path] = [];
+        }
+        acc[item.path].push(item);
+      } else {
+        acc[item.ref] = [item];
+      }
+    }
+    return acc;
+  }, [contents]);
+
+  const [grouping, setGrouping] = useState(true);
   const [searchText, setSearchText] = useState('');
 
   const total = contents.length ? contents.length - 1 : 0;
 
-  const height: number = clamp(95 + contents.length * 37, 132, 650);
+  const minHeight = 126;
+  const maxHeight = 658;
+  const headerHeight = 88;
+  const itemHeight = 38;
+  const height: number = clamp(
+    headerHeight +
+      (!grouping ? contents.length : Object.keys(contentsByPath).length) *
+        itemHeight,
+    minHeight,
+    maxHeight,
+  );
 
   return (
     <Window width={300} height={height} title={`Contents: ${total}`}>
@@ -36,13 +65,12 @@ export function LootPanel(props) {
         }}
       >
         <Section
+          scrollable={height === maxHeight}
           fill
-          scrollable
           title={
             <Stack>
               <Stack.Item grow>
                 <Input
-                  autoFocus
                   fluid
                   onInput={(event, value) => setSearchText(value)}
                   placeholder="Search"
@@ -68,7 +96,10 @@ export function LootPanel(props) {
           }
         >
           {grouping ? (
-            <GroupedContents contents={contents} searchText={searchText} />
+            <GroupedContents
+              contents={contentsByPath}
+              searchText={searchText}
+            />
           ) : (
             <RawContents contents={contents} searchText={searchText} />
           )}
